fix: moved mock data to testMockData folder
